+++ crates/zk-por-core/src/lib.rs
@@ -7,5 +7,5 @@ pub mod error;
 pub mod merkle_sum_prover;
 pub mod merkle_sum_tree;
 pub mod parser;
-pub mod recursive;
+pub mod recursive_prover;
 pub mod types;
